UNFINISHED: Refactored full assemble run without debugging to work properly in new fs
 - TODO: Use .then() to properly chain commands (things are failing because of promises)
